+++ source/Delphi.Forms.Info.pas
@@ -3,7 +3,7 @@ unit Delphi.Forms.Info;
 interface
 
 const
-  Version = '0.8.48.0';
+  Version = '0.8.49.0';
 
 implementation